backend/python/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
ftfy==6.1.1
regex==2023.5.5
tqdm==4.65.0
Pillow==9.5.0
requests==2.31.0
transformers
torch
open_clip_torch
python-multipart